Make ChordCollection an extension of a dictionary rather than a thing that has a dictionary
Having ChordCollection *have* a dict instead of *being* a dict was
done back in f7ca6fea as a part of an approach of storing the Chord
objects in the dictionary. That was handy because the objects could be
compared with each other for equality.

In f7ca6fea we reversed that decision by normalizing the chord names
as they were stored, which meant comparison still worked.

That means we can also go back to ChordCollection extending a dict,
which simplifies our implementation.