tests: claim the uke_config fixture is a UkeConfig to enable type checking
While this will in reality be a UkeTestConfig at run-time (as provided
by a pytest fixture), we don't care what happens at run time, only at
type-checking time, in which case a UkeConfig is actually more accurate.

Eventually we may simplify UkeConfig to be more of a dumb
dataclass (especially after separating out command-line logic) which
could actually be used for test configuration.